Changed to intuitive casting, removed dtype tests from basic test class
The dtype comparison was removed as the basic tests works with numpy, which does safe casting instead of intuitive casting.

+++ heat/core/tests/test_rounding.py
@@ -61,7 +61,7 @@ class TestRounding(TestCase):
         # for fabs
         self.assertEqual(int8_absolute_values_fabs.dtype, ht.float32)
         self.assertEqual(int16_absolute_values_fabs.dtype, ht.float32)
-        self.assertEqual(int32_absolute_values_fabs.dtype, ht.float64)
+        self.assertEqual(int32_absolute_values_fabs.dtype, ht.float32)
         self.assertEqual(int64_absolute_values_fabs.dtype, ht.float64)
         self.assertEqual(float32_absolute_values_fabs.dtype, ht.float32)
         self.assertEqual(float64_absolute_values_fabs.dtype, ht.float64)
