+++ heat/core/tests/test_trigonometrics.py
@@ -33,7 +33,7 @@ class TestTrigonometrics(TestCase):
         int32_tensor = ht.arange(elements, dtype=ht.int32)
         int32_arctan = ht.arctan(int32_tensor)
         self.assertIsInstance(int32_arctan, ht.DNDarray)
-        self.assertEqual(int32_arctan.dtype, ht.float64)
+        self.assertEqual(int32_arctan.dtype, ht.float32)
         self.assertTrue(torch.allclose(float32_arctan._DNDarray__array.double(), comparison))
 
         # arctan of longs, automatic conversion to intermediate floats
@@ -242,7 +242,7 @@ class TestTrigonometrics(TestCase):
         int32_tensor = ht.arange(elements, dtype=ht.int32)
         int32_cos = ht.cos(int32_tensor)
         self.assertIsInstance(int32_cos, ht.DNDarray)
-        self.assertEqual(int32_cos.dtype, ht.float64)
+        self.assertEqual(int32_cos.dtype, ht.float32)
         self.assertTrue(torch.allclose(float32_cos._DNDarray__array.double(), comparison))
 
         # cosine of longs, automatic conversion to intermediate floats
@@ -283,7 +283,7 @@ class TestTrigonometrics(TestCase):
         int32_tensor = ht.arange(elements, dtype=ht.int32)
         int32_cosh = ht.cosh(int32_tensor)
         self.assertIsInstance(int32_cosh, ht.DNDarray)
-        self.assertEqual(int32_cosh.dtype, ht.float64)
+        self.assertEqual(int32_cosh.dtype, ht.float32)
         self.assertTrue(torch.allclose(float32_cosh._DNDarray__array.double(), comparison))
 
         # hyperbolic cosine of longs, automatic conversion to intermediate floats
@@ -382,7 +382,7 @@ class TestTrigonometrics(TestCase):
         int32_tensor = ht.arange(elements, dtype=ht.int32)
         int32_sin = ht.sin(int32_tensor)
         self.assertIsInstance(int32_sin, ht.DNDarray)
-        self.assertEqual(int32_sin.dtype, ht.float64)
+        self.assertEqual(int32_sin.dtype, ht.float32)
         self.assertTrue(torch.allclose(int32_sin._DNDarray__array.double(), comparison))
 
         # sine of longs, automatic conversion to intermediate floats
@@ -423,7 +423,7 @@ class TestTrigonometrics(TestCase):
         int32_tensor = ht.arange(elements, dtype=ht.int32)
         int32_sinh = ht.sinh(int32_tensor)
         self.assertIsInstance(int32_sinh, ht.DNDarray)
-        self.assertEqual(int32_sinh.dtype, ht.float64)
+        self.assertEqual(int32_sinh.dtype, ht.float32)
         self.assertTrue(torch.allclose(int32_sinh._DNDarray__array.double(), comparison))
 
         # hyperbolic sine of longs, automatic conversion to intermediate floats
@@ -464,7 +464,7 @@ class TestTrigonometrics(TestCase):
         int32_tensor = ht.arange(elements, dtype=ht.int32)
         int32_tan = ht.tan(int32_tensor)
         self.assertIsInstance(int32_tan, ht.DNDarray)
-        self.assertEqual(int32_tan.dtype, ht.float64)
+        self.assertEqual(int32_tan.dtype, ht.float32)
         self.assertTrue(torch.allclose(int32_tan._DNDarray__array.double(), comparison))
 
         # tangent of longs, automatic conversion to intermediate floats
@@ -505,7 +505,7 @@ class TestTrigonometrics(TestCase):
         int32_tensor = ht.arange(elements, dtype=ht.int32)
         int32_tanh = ht.tanh(int32_tensor)
         self.assertIsInstance(int32_tanh, ht.DNDarray)
-        self.assertEqual(int32_tanh.dtype, ht.float64)
+        self.assertEqual(int32_tanh.dtype, ht.float32)
         self.assertTrue(torch.allclose(int32_tanh._DNDarray__array.double(), comparison))
 
         # hyperbolic tangent of longs, automatic conversion to intermediate floats
